requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.4.0
google-generativeai>=0.3.1
streamlit>=1.28.0
pymongo>=4.5.0
pytest>=7.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
plotly>=5.17.0
pandas>=2.0.0